requirements.txt
pandas
trendspy-lite
PyYAML
requests
python-dotenv
markdown
Pillow
google-generativeai